refactor: consolidate block toolbars into main header bar
This change moves all toolbar content from the secondary toolbar row
into the main header bar, eliminating the need for block-specific
TopBar.tsx files.

Changes:
- Move SplitToolbarLeft/Right portal mounts into SplitHeader
- Remove SplitToolbar component from SplitContainer
- Inline TopBar content into each block's main component
- Delete all block-specific TopBar.tsx files

The toolbar buttons (file menu, share, references, etc.) now render
immediately to the left of the title in the header bar.
