+++ js/app/packages/block-canvas/signal/topBarSignals.ts
@@ -1,0 +1,3 @@
+import { createBlockSignal } from '@core/block';
+
+export const connectorTypeMenuTriggerSignal = createBlockSignal(false);
