+++ js/app/packages/block-code/component/Block.tsx
@@ -1,9 +1,108 @@
-import { useIsNestedBlock } from '@core/block';
+import {
+  type FileOperation,
+  SplitFileMenu,
+} from '@app/component/split-layout/components/SplitFileMenu';
+import { SplitHeaderLeft } from '@app/component/split-layout/components/SplitHeader';
+import {
+  BlockItemSplitLabel,
+  SplitPermissionsBadge,
+} from '@app/component/split-layout/components/SplitLabel';
+import {
+  SplitToolbarLeft,
+  SplitToolbarRight,
+} from '@app/component/split-layout/components/SplitToolbar';
+import { withAnalytics } from '@coparse/analytics';
+import { useBlockId, useIsNestedBlock } from '@core/block';
 import { DocumentBlockContainer } from '@core/component/DocumentBlockContainer';
+import { DocumentPropertiesModal } from '@core/component/DocumentPropertiesModal';
+import { ReferencesModal } from '@core/component/ReferencesModal';
+import { ShareButton } from '@core/component/TopBar/ShareButton';
+import { blockMetadataSignal, blockTextSignal } from '@core/signal/load';
+import { useGetPermissions } from '@core/signal/permissions';
+import {
+  useBlockDocumentDownloadName,
+  useBlockDocumentName,
+} from '@core/util/currentBlockDocumentName';
+import { downloadFile } from '@filesystem/download';
+import Download from '@icon/regular/download-simple.svg';
+import { createCallback } from '@solid-primitives/rootless';
 import { Show } from 'solid-js';
 import { CodeMarkdown } from './CodeMarkdown';
 import { CodeMirror } from './CodeMirror';
-import { TopBar } from './TopBar';
+
+const { track, TrackingEvents } = withAnalytics();
+
+function CodeTopBar() {
+  const blockId = useBlockId();
+  const text = blockTextSignal.get;
+  const name = useBlockDocumentName();
+  const downloadName = useBlockDocumentDownloadName();
+  const userPermissions = useGetPermissions();
+
+  const downloadDocument = createCallback(() => {
+    const content = text();
+    if (!text || !name) return;
+    const file = new Blob([content ?? ''], { type: 'text/plain' });
+    downloadFile(file, downloadName());
+    track(TrackingEvents.BLOCKCODE.FILEMENU.DOWNLOAD);
+  });
+
+  const ops: FileOperation[] = [
+    { op: 'pin' },
+    { op: 'rename' },
+    { op: 'copy' },
+    { op: 'moveToProject' },
+    {
+      label: 'Download',
+      icon: Download,
+      action: downloadDocument,
+      divideAbove: true,
+    },
+    { op: 'delete', divideAbove: true },
+  ];
+
+  return (
+    <>
+      <SplitHeaderLeft>
+        <BlockItemSplitLabel />
+      </SplitHeaderLeft>
+      <SplitToolbarLeft>
+        <div class="p-1">
+          <SplitFileMenu
+            id={blockId}
+            itemType="document"
+            name={name()}
+            ops={ops}
+          />
+        </div>
+      </SplitToolbarLeft>
+      <SplitToolbarRight>
+        <div class="flex items-center p-1">
+          <ReferencesModal
+            documentId={blockId}
+            documentName={name()}
+            buttonSize="sm"
+          />
+          <DocumentPropertiesModal
+            documentId={blockId}
+            blockType="code"
+            buttonSize="sm"
+          />
+          <div class="flex items-center">
+            <SplitPermissionsBadge />
+            <ShareButton
+              id={blockId}
+              name={name()}
+              userPermissions={userPermissions()}
+              itemType="document"
+              owner={blockMetadataSignal()?.owner}
+            />
+          </div>
+        </div>
+      </SplitToolbarRight>
+    </>
+  );
+}
 
 export default function BlockCode() {
   const isNestedBlock = useIsNestedBlock();
@@ -12,7 +111,7 @@ export default function BlockCode() {
     <DocumentBlockContainer usesCenterBar>
       <Show when={!isNestedBlock} fallback={<CodeMarkdown />}>
         <div class="size-full bg-panel select-none overscroll-none overflow-hidden flex flex-col items-end relative">
-          <TopBar />
+          <CodeTopBar />
           <CodeMirror />
         </div>
       </Show>
