+++ js/app/packages/block-project/component/Block.tsx
@@ -1,6 +1,19 @@
 import { useGlobalBlockOrchestrator } from '@app/component/GlobalAppState';
 import { PreviewPanel } from '@app/component/PreviewPanel';
 import { SplitPanelContext } from '@app/component/split-layout/context';
+import {
+  type FileOperation,
+  SplitFileMenu,
+} from '@app/component/split-layout/components/SplitFileMenu';
+import { SplitHeaderLeft } from '@app/component/split-layout/components/SplitHeader';
+import {
+  BlockItemSplitLabel,
+  SplitPermissionsBadge,
+} from '@app/component/split-layout/components/SplitLabel';
+import {
+  SplitToolbarLeft,
+  SplitToolbarRight,
+} from '@app/component/split-layout/components/SplitToolbar';
 import { useSplitPanelOrThrow } from '@app/component/split-layout/layoutUtils';
 import { UnifiedListView } from '@app/component/UnifiedListView';
 import { PROJECT_VIEWCONFIG_BASE } from '@app/component/ViewConfig';
@@ -9,16 +22,26 @@ import { getIsSpecialProject } from '@block-project/isSpecial';
 import { useBlockId } from '@core/block';
 import { DocumentBlockContainer } from '@core/component/DocumentBlockContainer';
 import { FileDropOverlay } from '@core/component/FileDropOverlay';
-import { ENABLE_PROJECT_VIEW_PREVIEW } from '@core/constant/featureFlags';
+import { ShareButton } from '@core/component/TopBar/ShareButton';
+import {
+  ENABLE_PROJECT_SHARING,
+  ENABLE_PROJECT_VIEW_PREVIEW,
+} from '@core/constant/featureFlags';
 import { fileFolderDrop } from '@core/directive/fileFolderDrop';
 import { fileSelector } from '@core/directive/fileSelector';
 import { registerHotkey } from '@core/hotkey/hotkeys';
 import { TOKENS } from '@core/hotkey/tokens';
 import {
+  useCanEdit,
+  useGetPermissions,
+  useIsDocumentOwner,
+} from '@core/signal/permissions';
+import {
   handleFileFolderDrop,
   type UploadInput,
   uploadFiles,
 } from '@core/util/upload';
+import { buildSimpleEntityUrl } from '@core/util/url';
 import {
   queryKeys,
   useQueryClient as useEntityQueryClient,
@@ -35,11 +58,97 @@ import {
   untrack,
 } from 'solid-js';
 import { projectBlockDataSignal } from '../signal/projectBlockData';
-import { TopBar } from './TopBar';
+import { ProjectCreateMenu } from './ProjectCreateMenu';
+import { ProjectPropertiesModal } from './ProjectPropertiesModal';
 
 // HACK: prevent lint error on custom directive
 false && fileFolderDrop;
 false && fileSelector;
+
+function ProjectTopBar() {
+  const splitPanelContext = useSplitPanelOrThrow();
+  const [preview] = splitPanelContext.previewState;
+  const id = useBlockId();
+  const isSpecialProject = getIsSpecialProject(id);
+  const permissions = useGetPermissions();
+  const isOwner = useIsDocumentOwner();
+  const canEdit = useCanEdit();
+  const name = () => projectBlockDataSignal()?.projectMetadata.name ?? '';
+  const owner = () => projectBlockDataSignal()?.projectMetadata.userId;
+
+  function handleCopyLink() {
+    navigator.clipboard.writeText(
+      buildSimpleEntityUrl(
+        {
+          type: 'project',
+          id,
+        },
+        {}
+      )
+    );
+    toast.success('Link copied to clipboard');
+  }
+
+  const ops = createMemo<FileOperation[]>(() => [
+    ...(isOwner() && !isSpecialProject
+      ? [
+          { op: 'rename' as const },
+          { op: 'moveToProject' as const },
+          { op: 'delete' as const, divideAbove: true },
+        ]
+      : []),
+  ]);
+
+  const showToolbarRight = () => {
+    if (!ENABLE_PROJECT_VIEW_PREVIEW) return true;
+    return !preview();
+  };
+
+  return (
+    <>
+      <SplitHeaderLeft>
+        <BlockItemSplitLabel fallbackName={name()} />
+      </SplitHeaderLeft>
+      <SplitToolbarLeft class="flex-0">
+        <div class="flex gap-2 p-1">
+          <Show when={ops().length > 0}>
+            <SplitFileMenu
+              id={id}
+              itemType="project"
+              name={name()}
+              ops={ops()}
+            />
+            <Show when={canEdit()}>
+              <ProjectCreateMenu id={id} />
+            </Show>
+          </Show>
+        </div>
+      </SplitToolbarLeft>
+      <Show when={showToolbarRight()}>
+        <SplitToolbarRight>
+          <div class="flex items-center p-1">
+            <div class="flex items-center">
+              <Show when={!isSpecialProject}>
+                <ProjectPropertiesModal buttonSize="sm" name={name()} />
+              </Show>
+              <SplitPermissionsBadge />
+              <Show when={ENABLE_PROJECT_SHARING && !isSpecialProject}>
+                <ShareButton
+                  id={id}
+                  name={name()}
+                  userPermissions={permissions()}
+                  copyLink={handleCopyLink}
+                  itemType="project"
+                  owner={owner()}
+                />
+              </Show>
+            </div>
+          </div>
+        </SplitToolbarRight>
+      </Show>
+    </>
+  );
+}
 
 const Block: Component = () => {
   const [isDragging, setIsDragging] = createSignal(false);
@@ -157,7 +266,7 @@ const Block: Component = () => {
         <Show when={isDragging() && !isSpecialProject}>
           <FileDropOverlay>Upload to this folder</FileDropOverlay>
         </Show>
-        <TopBar />
+        <ProjectTopBar />
         <Show when={ENABLE_PROJECT_VIEW_PREVIEW} fallback={<UnifiedListView />}>
           <div class="flex size-full">
             <SplitPanelContext.Provider
