--- js/app/packages/block-unknown/component/TopBar.tsx
@@ -1,96 +1,0 @@
-import {
-  type FileOperation,
-  SplitFileMenu,
-} from '@app/component/split-layout/components/SplitFileMenu';
-import { SplitHeaderLeft } from '@app/component/split-layout/components/SplitHeader';
-import {
-  BlockItemSplitLabel,
-  SplitPermissionsBadge,
-} from '@app/component/split-layout/components/SplitLabel';
-import {
-  SplitToolbarLeft,
-  SplitToolbarRight,
-} from '@app/component/split-layout/components/SplitToolbar';
-import { useBlockId } from '@core/block';
-import { ReferencesModal } from '@core/component/ReferencesModal';
-import { ShareButton } from '@core/component/TopBar/ShareButton';
-import { blockMetadataSignal } from '@core/signal/load';
-import { useGetPermissions } from '@core/signal/permissions';
-import {
-  useBlockDocumentDownloadName,
-  useBlockDocumentName,
-} from '@core/util/currentBlockDocumentName';
-import { downloadFile } from '@filesystem/download';
-import DownloadSimple from '@icon/regular/download-simple.svg';
-import { createCallback } from '@solid-primitives/rootless';
-import { toast } from 'core/component/Toast/Toast';
-import { useGetFileBlob } from '../signal/blockData';
-
-export function TopBar() {
-  const blockId = useBlockId();
-  const fileName = useBlockDocumentName();
-  const downloadName = useBlockDocumentDownloadName();
-  const getBlob = useGetFileBlob();
-  const userPermissions = useGetPermissions();
-
-  const downloadDocument = createCallback(async () => {
-    try {
-      const blob = await getBlob();
-      downloadFile(blob, downloadName());
-    } catch (e) {
-      console.error('error downloading file', e);
-      toast.failure('Error downloading file');
-    }
-  });
-
-  const ops: FileOperation[] = [
-    { op: 'pin' },
-    { op: 'rename' },
-    { op: 'copy' },
-    { op: 'moveToProject' },
-    {
-      label: 'Download',
-      icon: DownloadSimple,
-      action: downloadDocument,
-      divideAbove: true,
-    },
-    { op: 'delete', divideAbove: true },
-  ];
-
-  return (
-    <>
-      <SplitHeaderLeft>
-        <BlockItemSplitLabel />
-      </SplitHeaderLeft>
-      <SplitToolbarLeft>
-        <div class="p-1">
-          <SplitFileMenu
-            id={blockId}
-            itemType="document"
-            name={fileName()}
-            ops={ops}
-          />
-        </div>
-      </SplitToolbarLeft>
-      <SplitToolbarRight>
-        <div class="flex items-center p-1">
-          <ReferencesModal
-            documentId={blockId}
-            documentName={fileName()}
-            buttonSize="sm"
-          />
-          <div class="flex items-center">
-            <SplitPermissionsBadge />
-            <ShareButton
-              id={blockId}
-              name={fileName()}
-              userPermissions={userPermissions()}
-              itemType="document"
-              owner={blockMetadataSignal()?.owner}
-            />
-          </div>
-        </div>
-      </SplitToolbarRight>
-    </>
-  );
-}
